Migration to OpenTelemetry
This inserts the wiring for opentelemetry-collector into both the oc
and otel bindings, and inserts the opentelemetry instrumentation into
the places which it immediately supports.

Existing uses of OpenCensus are updated to OpenTelemetry. The OC
exporters are retained for libraries which still use OC (notably
gocloud.dev), and contexts are mapped as needed.

Direct support for jaeger and stackdriver is dropped in favor of
configuring this in opentelemtry-collector, which supports far more
backends. I'll follow up with examples in bb-deployments.

A matching PR to bb-remote-execution will need to be merged along with
this - there turns out to be no neat way to do the transition.